[kahi_scholar_works] implemented support to process works without doi using similarity (issue #245)
